Module wiggle_generate::config

Structs

AsyncConf

Modules and funcs that have async signatures

AsyncConfField

Config

ErrorConf

Map from abi error type to rich error type

ErrorConfField

Literal

A single witx document, provided as a string literal.

Paths

A collection of paths, pointing to witx documents.

WasmtimeConfig

Enums

AsyncFunctions

Asyncness

ConfigField

WasmtimeConfigField

WitxConf

The witx document(s) that will be loaded from a Config.