Struct wiggle_generate::CodegenSettings

pub struct CodegenSettings {
    pub errors: ErrorTransform,
    pub async_: AsyncConf,
    pub wasmtime: bool,
    pub tracing: TracingConf,
    pub mutable: bool,
}

Fields
errors: ErrorTransformasync_: AsyncConfwasmtime: booltracing: TracingConf

Disabling this feature makes it possible to remove all of the tracing code emitted in the Wiggle-generated code; this can be helpful while inspecting the code (e.g., with cargo expand).

mutable: bool

Determine whether the context structure will use &mut self (true) or simply &self.

Implementations

impl CodegenSettings

pub fn new(
error_conf: &ErrorConf,
async_: &AsyncConf,
doc: &Document,
wasmtime: bool,
tracing: &TracingConf,
mutable: bool
) -> Result<Self, Error>

pub fn get_async(&self, module: &Module, func: &InterfaceFunc) -> Asyncness