Module wiggle_generate::names

Functions

builtin_type

enum_variant

Convert an enum variant from its Id name to its Rust Ident representation.

flag_member

func

Convert a function name from its Id name to its Rust Ident representation.

func_len_binding

For when you need a {name}_len binding for passing an array:

func_param

Convert a parameter name from its Id name to its Rust Ident representation.

func_ptr_binding

For when you need a {name}_ptr binding for passing a value by reference:

int_member

module

Convert a module name from its Id name to its Rust Ident representation.

struct_member

Convert a struct member from its Id name to its Rust Ident representation.

trait_name

Convert a trait name from its Id name to its Rust Ident representation.

type_

type_ref

user_error_conversion_method

wasm_type